move out filter to its own file
